refactor view context, view modules
