+++ webportal/signals.py
@@ -1,5 +1,6 @@
 from django.db.models.signals import post_migrate
 from django.dispatch import receiver
+
 from .models import Team
